chore: move bigquery/table.py to bigquery/_operations/table.py
